Persona y cliente con insert y update

+++ src/main/java/com/cuenta/bancaria/cuenta/bancaria/controller/ClienteController.java
@@ -4,7 +4,10 @@
 package com.cuenta.bancaria.cuenta.bancaria.controller;
 
 import java.util.List;
+import java.util.Optional;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -18,8 +21,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.cuenta.bancaria.cuenta.bancaria.controller.dto.ClienteEntradaDto;
 import com.cuenta.bancaria.cuenta.bancaria.model.Cliente;
+import com.cuenta.bancaria.cuenta.bancaria.model.Persona;
 import com.cuenta.bancaria.cuenta.bancaria.service.ClienteService;
+import com.cuenta.bancaria.cuenta.bancaria.service.PersonaService;
 
 /**
  * @author JAVIM
@@ -28,33 +34,90 @@ import com.cuenta.bancaria.cuenta.bancaria.service.ClienteService;
 @RestController
 @RequestMapping("/api/clientes")
 public class ClienteController {
-	
+	private static final Logger log = LoggerFactory.getLogger(ClienteController.class);
+
 	@Autowired
 	private ClienteService service;
-	
-	
+
+	@Autowired
+	private PersonaService personaService;
+
 	@PostMapping
-	public ResponseEntity<Cliente> create(@Validated @RequestBody Cliente cliente) {
-		Cliente clienteGuardado = service.create(cliente);
-		return new ResponseEntity<Cliente>(clienteGuardado, HttpStatus.CREATED);
+	public ResponseEntity<?> create(@Validated @RequestBody ClienteEntradaDto clienteEntradaDto) {
+		try {
+			Optional<Persona> personaEncontrada = personaService
+					.obtenerPorIdentificacion(clienteEntradaDto.getIdentificacion());
+			if (!personaEncontrada.isPresent()) {
+				return new ResponseEntity<>("No existe la persona con esta identificación", HttpStatus.BAD_REQUEST);
+			}
+			Optional<Cliente> clienteEncontrado = service.obtenerPorEstadoIdCliente(Boolean.TRUE.toString(),
+					personaEncontrada.get().getIdPersona());
+			if (clienteEncontrado.isEmpty()) {
+				Cliente cliente = new Cliente();
+				cliente.setContrasena(clienteEntradaDto.getContrasenia());
+				cliente.setEstado(Boolean.TRUE.toString());
+				cliente.setPersona(personaEncontrada.get());
+				Cliente clienteGuardado = service.create(cliente);
+				return new ResponseEntity<Cliente>(clienteGuardado, HttpStatus.CREATED);
+			} else {
+				return new ResponseEntity<>("Cliente ya se encuentra registrado", HttpStatus.CREATED);
+			}
+
+		} catch (Exception e) {
+			log.error("Por favor comuniquese con el administrador", e);
+			return new ResponseEntity<>("Por favor comuniquese con el administrador", HttpStatus.INTERNAL_SERVER_ERROR);
+		}
 	}
-	
+
 	@GetMapping
 	public ResponseEntity<List<Cliente>> read() {
 		List<Cliente> list = service.read();
 		return new ResponseEntity<List<Cliente>>(list, HttpStatus.OK);
+
 	}
-	
+
 	@PutMapping
-	public ResponseEntity<Cliente> update(@Validated @RequestBody Cliente employee) {
-		Cliente clienteGuardado = service.update(employee);
-		return new ResponseEntity<Cliente>(clienteGuardado, HttpStatus.OK);
+	public ResponseEntity<?> update(@Validated @RequestBody ClienteEntradaDto clienteEntradaDto) {
+		try {
+
+			Optional<Persona> personaEncontrada = personaService
+					.obtenerPorIdentificacion(clienteEntradaDto.getIdentificacion());
+			if (!personaEncontrada.isPresent()) {
+				return new ResponseEntity<>("No existe el cliente con esta identificación", HttpStatus.BAD_REQUEST);
+			}
+			Optional<Cliente> clienteEncontrado = service.obtenerPorEstadoIdCliente(Boolean.TRUE.toString(),
+					personaEncontrada.get().getIdPersona());
+			if (clienteEncontrado.isPresent()) {
+				Cliente cliente = clienteEncontrado.get();
+				cliente.setContrasena(clienteEntradaDto.getContrasenia());
+				cliente.setEstado(Boolean.TRUE.toString());
+				Cliente clienteGuardado = service.update(cliente);
+				return new ResponseEntity<Cliente>(clienteGuardado, HttpStatus.OK);
+			} else {
+				return new ResponseEntity<>("Cliente no se encuentra registrado", HttpStatus.CREATED);
+			}
+		} catch (Exception e) {
+			log.error("Por favor comuniquese con el administrador", e);
+			return new ResponseEntity<>("Por favor comuniquese con el administrador", HttpStatus.INTERNAL_SERVER_ERROR);
+		}
 	}
 
 	@DeleteMapping("/{id}")
-	public ResponseEntity<Void> delete(@PathVariable("id") Long id) {
-		service.delete(id);
-		return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
+	public ResponseEntity<?> delete(@PathVariable("id") Long id) {
+		try {
+			
+			Optional<Cliente> personaEncontrada = service.obtenerPorId(id);
+			if (personaEncontrada.isPresent()) {
+				service.delete(id);
+				return new ResponseEntity<>("Registro Eliminado",HttpStatus.OK);
+			}
+			return new ResponseEntity<String>(
+					"No se puede eliminar el registro con id: " + id + " no existe el registro",
+					HttpStatus.BAD_REQUEST);
+		} catch (Exception e) {
+			log.error("Por favor comuniquese con el administrador", e);
+			return new ResponseEntity<>("Por favor comuniquese con el administrador", HttpStatus.INTERNAL_SERVER_ERROR);
+		}
 	}
 
 }
